+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - Make the broadcast function public on Sender
 - Added a `call_as` helper on TxHandler Trait
 - Added optional disable of sorting of ExecuteFns and QueryFns fields by @Kayanski
-
+- Added automatic tx retrying in case of an account sequence error by @Kayanski
+- Made transaction retrying more modular (for adding more retry cases) by Kayanski
 
 ## v0.15.0